Draft of extracted mapping logic.
