+++ README.md
@@ -83,20 +83,23 @@ num_spin_orbitals = molecule.num_orbitals * 2
 # Build the qubit operator, which is the input to the VQE algorithm in Aqua
 ferm_op = FermionicOperator(h1=molecule.one_body_integrals, h2=molecule.two_body_integrals)
 map_type = 'PARITY'
-qubit_op = ferm_op.mapping(map_type)
+qubit_op = mapping(map_type)
 qubit_op = Z2Symmetries.two_qubit_reduction(qubit_op, num_particles)
 num_qubits = qubit_op.num_qubits
 
 # setup a classical optimizer for VQE
 from qiskit.aqua.components.optimizers import L_BFGS_B
+
 optimizer = L_BFGS_B()
 
 # setup the initial state for the variational form
 from qiskit_nature.circuit.library import HartreeFock
+
 init_state = HartreeFock(num_spin_orbitals, num_particles)
 
 # setup the variational form for VQE
 from qiskit.circuit.library import TwoLocal
+
 var_form = TwoLocal(num_qubits, ['ry', 'rz'], 'cz')
 
 # add the initial state
@@ -104,10 +107,12 @@ var_form.compose(init_state, front=True)
 
 # setup and run VQE
 from qiskit.aqua.algorithms import VQE
+
 algorithm = VQE(qubit_op, var_form, optimizer)
 
 # set the backend for the quantum computation
 from qiskit import Aer
+
 backend = Aer.get_backend('statevector_simulator')
 
 result = algorithm.run(backend)
